Remove compilation error when using clang 8:
error: non-constant-expression cannot be narrowed from type 'int' to 'CGFloat'

git-svn-id: file:///fltk/svn/fltk/branches/branch-1.4@13070 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_cocoa.mm
@@ -3922,7 +3922,7 @@ int Fl_Cocoa_Screen_Driver::dnd(int use_selection)
     NSPasteboardItem *pbItem = [[[NSPasteboardItem alloc] init] autorelease];
     [pbItem setData:(NSData*)text forType:UTF8_pasteboard_type];
     NSDraggingItem *dragItem = [[[NSDraggingItem alloc] initWithPasteboardWriter:pbItem] autorelease];
-    NSRect r = {pt, {width, height}};
+    NSRect r = {pt, {CGFloat(width), CGFloat(height)}};
     [dragItem setDraggingFrame:r contents:image];
     [myview beginDraggingSessionWithItems:[NSArray arrayWithObject:dragItem] event:theEvent source:myview];
   } else
